refactor: remove dead useChatStore, move getFeatureModelConfig to appStore
- Delete unused chatStore.ts (230 lines of dead Pinia store code)
- Move getFeatureModelConfig to appStore.ts as standalone export
- Add storeApi.delete('chatStore') cleanup in fetchLlmSettings
- Update AGENTS.md and docs references
- No functional changes — store was never instantiated
