+++ src/store/index.ts
@@ -2,7 +2,6 @@ export * from './appStore';
 export * from './userStore';
 export * from './connectionStore';
 export * from './fileStore.ts';
-export * from './chatStore';
 export * from './clusterManageStore';
 export * from './dynamoManageStore.ts';
 export * from './importExportStore.ts';
